refactor(api): consolidate uniform agent-target clan/faction tools (#22)
SonarCloud Automatic Analysis ignores sonar-project.properties, so the duplication quality
gate must be satisfied by reducing real duplication rather than excluding the tool layer.

The flagged blocks were the structurally-identical single-agent-target verbs (CPD tokenizes
the differing @Tool strings uniformly). Extract a shared submitAgentTargeted helper and group
the verbs into ClanMembershipTools (kick/promote/demote/transfer) and FactionRankTools
(promote/demote), so each verb is a short method below CPD's block threshold and the repeated
ctor/annotation/body blocks are gone. Behaviour and @Tool names/descriptions are unchanged.

Removes the now-non-functional sonar-project.properties.
