+++ api/src/main/kotlin/dev/gvart/genesara/api/internal/mcp/McpServerConfiguration.kt
@@ -10,21 +10,17 @@ import dev.gvart.genesara.api.internal.mcp.tools.attributes.AllocatePointsTool
 import dev.gvart.genesara.api.internal.mcp.tools.build.BuildTool
 import dev.gvart.genesara.api.internal.mcp.tools.chest.DepositToChestTool
 import dev.gvart.genesara.api.internal.mcp.tools.chest.WithdrawFromChestTool
+import dev.gvart.genesara.api.internal.mcp.tools.clan.ClanMembershipTools
 import dev.gvart.genesara.api.internal.mcp.tools.clan.CreateClanTool
-import dev.gvart.genesara.api.internal.mcp.tools.clan.DemoteClanMemberTool
 import dev.gvart.genesara.api.internal.mcp.tools.clan.DissolveClanTool
 import dev.gvart.genesara.api.internal.mcp.tools.clan.GetClanStatusTool
 import dev.gvart.genesara.api.internal.mcp.tools.clan.InviteToClanTool
-import dev.gvart.genesara.api.internal.mcp.tools.clan.KickClanMemberTool
 import dev.gvart.genesara.api.internal.mcp.tools.clan.LeaveClanTool
-import dev.gvart.genesara.api.internal.mcp.tools.clan.PromoteClanMemberTool
 import dev.gvart.genesara.api.internal.mcp.tools.clan.RespondClanInviteTool
-import dev.gvart.genesara.api.internal.mcp.tools.clan.TransferClanLeadershipTool
 import dev.gvart.genesara.api.internal.mcp.tools.faction.CreateFactionTool
-import dev.gvart.genesara.api.internal.mcp.tools.faction.DemoteFactionMemberTool
+import dev.gvart.genesara.api.internal.mcp.tools.faction.FactionRankTools
 import dev.gvart.genesara.api.internal.mcp.tools.faction.InviteClanToFactionTool
 import dev.gvart.genesara.api.internal.mcp.tools.faction.LeaveFactionTool
-import dev.gvart.genesara.api.internal.mcp.tools.faction.PromoteFactionMemberTool
 import dev.gvart.genesara.api.internal.mcp.tools.faction.RespondFactionInviteTool
 import dev.gvart.genesara.api.internal.mcp.tools.classselect.SelectClassTool
 import dev.gvart.genesara.api.internal.mcp.tools.classselect.SelectEvolutionTool
@@ -134,19 +130,15 @@ internal class McpServerConfiguration {
         createClan: CreateClanTool,
         leaveClan: LeaveClanTool,
         dissolveClan: DissolveClanTool,
-        transferClanLeadership: TransferClanLeadershipTool,
         getClanStatus: GetClanStatusTool,
         inviteToClan: InviteToClanTool,
         respondClanInvite: RespondClanInviteTool,
-        kickClanMember: KickClanMemberTool,
-        promoteClanMember: PromoteClanMemberTool,
-        demoteClanMember: DemoteClanMemberTool,
+        clanMembership: ClanMembershipTools,
         createFaction: CreateFactionTool,
         inviteClanToFaction: InviteClanToFactionTool,
         respondFactionInvite: RespondFactionInviteTool,
         leaveFaction: LeaveFactionTool,
-        promoteFactionMember: PromoteFactionMemberTool,
-        demoteFactionMember: DemoteFactionMemberTool,
+        factionRank: FactionRankTools,
         getEvents: GetEventsTool,
     ): ToolCallbackProvider {
         val methodProvider = MethodToolCallbackProvider.builder()
@@ -164,10 +156,9 @@ internal class McpServerConfiguration {
                 tame, mountTransport, dismountTransport, maintain,
                 storeOnMount, takeFromMount,
                 partyInvite, partyRespond, leaveParty, kickMember, getParty,
-                createClan, leaveClan, dissolveClan, transferClanLeadership, getClanStatus,
-                inviteToClan, respondClanInvite, kickClanMember, promoteClanMember, demoteClanMember,
-                createFaction, inviteClanToFaction, respondFactionInvite, leaveFaction,
-                promoteFactionMember, demoteFactionMember,
+                createClan, leaveClan, dissolveClan, getClanStatus,
+                inviteToClan, respondClanInvite, clanMembership,
+                createFaction, inviteClanToFaction, respondFactionInvite, leaveFaction, factionRank,
                 getEvents,
             )
             .build()
