--- sonar-project.properties
@@ -1,13 +1,0 @@
-# SonarCloud analysis configuration (read by Automatic Analysis).
-
-# The MCP tool layer is intentionally uniform: one @Component @Tool class per agent verb,
-# differing only by the @Tool name/description and the WorldCommand it builds. SonarCloud's
-# token-based copy-paste detector treats every string literal as one token, so it flags each
-# new tool as a duplicate of the ~50 existing ones — noise, not actionable duplication.
-# Exclude the thin adapter layer from CPD (it stays in scope for bug/smell analysis).
-sonar.cpd.exclusions=**/internal/mcp/tools/**/*Tool.kt
-
-# Flyway DDL migrations are append-only Postgres SQL, reviewed in PRs. SonarCloud applies
-# Oracle-flavoured SQL rules to them (e.g. "use VARCHAR2 instead of VARCHAR"), which are wrong
-# for Postgres. Keep them out of analysis rather than carry false-positive smells.
-sonar.exclusions=**/db/migration/**
